# docs/swagger.yaml
basePath: /api
definitions:
...
  friendship.SendRequestRequest:
    properties:
      addresseeId:
        type: string
    required:
    - addresseeId
    type: object
...
  models.Friendship:
    properties:
      addresseeId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      requesterId:
        type: string
      respondedAt:
        type: string
      status:
        $ref: '#/definitions/models.FriendshipStatus'
    type: object
  models.FriendshipStatus:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - FriendshipPending
    - FriendshipAccepted
...
  models.Reward:
    properties:
      cost:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      isAvailable:
        type: boolean
      redeemedById:
        type: string
      rewardGiverId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      visibility:
        $ref: '#/definitions/models.RewardVisibilityMode'
    type: object
  models.RewardVisibilityMode:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Public
    - Restricted
  models.Transaction:
    properties:
      amount:
        type: integer
      giverId:
        type: string
      id:
        type: string
      referenceId:
        type: string
      timestamp:
        type: string
      type:
        $ref: '#/definitions/models.TransactionType'
      userId:
        type: string
    type: object
  models.TransactionType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Earned
    - Spent
  models.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  models.UserPointBalance:
    properties:
      balance:
        type: integer
      giverId:
        type: string
      ownerId:
        type: string
    type: object
  reward.CreateRewardRequest:
    properties:
      allowedUserIds:
        items:
          type: string
        type: array
      cost:
        type: integer
      description:
        type: string
      title:
        type: string
      visibility:
        $ref: '#/definitions/models.RewardVisibilityMode'
    required:
    - cost
    - title
    type: object
...
info:
  contact: {}
  description: A health and longevity challenge API
  title: Panjang Umur API
  version: "1.0"
paths:
...
  /friends:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List my friends
      tags:
      - Friends
  /friends/{userId}:
    delete:
      parameters:
      - description: Friend's user ID (UUID)
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a friend
      tags:
      - Friends
  /friends/requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List my incoming and outgoing friend requests
      tags:
      - Friends
    post:
      consumes:
      - application/json
      parameters:
      - description: Addressee
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friendship.SendRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Friendship'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send a friend request
      tags:
      - Friends
  /friends/requests/{id}/accept:
    post:
      parameters:
      - description: Friend request ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Friendship'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Accept a friend request
      tags:
      - Friends
  /friends/requests/{id}/decline:
    post:
      parameters:
      - description: Friend request ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Decline or cancel a friend request
      tags:
      - Friends
  /rewards:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reward data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/reward.CreateRewardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Reward'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a reward in my shop
      tags:
      - Rewards
  /rewards/{rewardId}/cancel:
    patch:
      parameters:
      - description: Reward ID (UUID)
        in: path
        name: rewardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Reward'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel an available reward
      tags:
      - Rewards
  /rewards/{rewardId}/redeem:
    patch:
      parameters:
      - description: Reward ID (UUID)
        in: path
        name: rewardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Reward'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Redeem a reward using points
      tags:
      - Rewards
  /rewards/me/given:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Reward'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List rewards I have created (my shop management view)
      tags:
      - Rewards
  /rewards/shop/{giverId}:
    get:
      parameters:
      - description: Reward giver's user ID (UUID)
        in: path
        name: giverId
        required: true
        type: string
      - description: Filter to available rewards only
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Reward'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 'Browse a friend''s shop (rewards visible to me: public + restricted-to-me)'
      tags:
      - Rewards
  /transactions/balances:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserPointBalance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my point balances per giver
      tags:
      - Transactions
  /transactions/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my transaction history
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    description: Use "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"